@@ -1,10 +1,11 @@
 {
   "private": true,
+  "type": "module",
   "dependencies": {
-    "@fastly/compute-js-static-publish": "^6.0.0",
+    "@fastly/compute-js-static-publish": "^6.1.0",
     "@fastly/js-compute": "^3.11.0",
     "@h7/compute-js-formdata": "^0.1.0",
-    "@h7/compute-js-rsc": "^0.1.0-alpha.0",
+    "@h7/compute-js-rsc": "^0.1.0-alpha.2",
     "blob-polyfill": "^7.0.20220408",
     "formdata-polyfill": "^4.0.0"
   },
@@ -12,16 +13,17 @@
     "@babel/core": "^7.23.2",
     "@babel/preset-react": "^7.22.15",
     "babel-loader": "^9.1.3",
-    "react": "18.3.0-canary-a4939017f-20240320",
-    "react-server-dom-webpack": "18.3.0-canary-a4939017f-20240320",
+    "react": "19.0.0-canary-a73c3450e-20240329",
+    "react-server-dom-webpack": "19.0.0-canary-a73c3450e-20240329",
     "webpack": "^5.89.0",
-    "webpack-cli": "^5.1.4"
+    "webpack-cli": "^5.1.4",
+    "webpack-manifest-plugin": "^5.0.0"
   },
   "scripts": {
     "clean": "rm -rf bin build pkg",
     "prebuild": "webpack && compute-js-static-publish --build-static --suppress-framework-warnings",
-    "build": "js-compute-runtime ./server/index.js ./bin/main.wasm",
-    "start": "fastly compute serve",
+    "build": "js-compute-runtime --enable-experimental-top-level-await ./server/index.js ./bin/main.wasm",
+    "start": "fastly compute serve --verbose",
     "deploy": "fastly compute publish"
   }
 }
