@@ -1,6 +1,6 @@
 const development = (process.env.NODE_ENV || "development") === "development";
 
-module.exports = {
+export default {
   presets: [
     [
       "@babel/preset-react",
