@@ -1,7 +1,3 @@
-const rscBackend = require('@h7/compute-js-rsc/backend');
-const { App } = require('./app/app.jsx');
-
-module.exports = {
-  rscBackend,
-  App,
-};
+export * as React from 'react';
+export * as rscBackend from '@h7/compute-js-rsc/backend';
+export { App } from './app/app.jsx';
